Fix remaining PHPStan findings with real typed getters, not baseline suppressions
Replaces getData()/mixed-typed access with real typed getters and array-shape
docblocks across the Cron jobs, approval/track controllers, the Flow editor
block, and DashboardViewModel. Fixes a genuine bug in DashboardViewModel::
getCampaigns() (was returning DataObject[] via getItems() instead of typed
Campaign entities) and gives TriggerEvent::toOptionArray() a proper return
shape, which was the root cause of several "offset access on mixed" findings
downstream in Flow.php.

Investigated and confirmed as genuine (left baselined, not hand-waved):
ReorderCycleRepository has no save() (read-only repo), Quote::load() has no
uniform guest+registered repository equivalent, and Zend_Db_Select vs Select
is a real Magento core stub inconsistency. The CampaignDispatcher
"property never read" baseline entry was checked directly against the code
and is stale — both factories are genuinely used — no code change needed.

+++ Cron/PruneVisitorEvents.php
@@ -26,7 +26,7 @@ class PruneVisitorEvents
     public function execute(): void
     {
         $retentionDays = $this->config->getTrackingRetentionDays();
-        $cutoff = date('Y-m-d H:i:s', strtotime("-{$retentionDays} days"));
+        $cutoff = date('Y-m-d H:i:s', (int) strtotime("-{$retentionDays} days"));
 
         $connection = $this->resourceConnection->getConnection();
         $table = $this->resourceConnection->getTableName('ordo_visitor_event');
